Extract events from VM state file
