+++ ui/src/main/kotlin/com/processout/sdk/ui/card/tokenization/v2/CardTokenizationViewModelState.kt
@@ -1,9 +1,6 @@
 package com.processout.sdk.ui.card.tokenization.v2
 
 import androidx.compose.runtime.Immutable
-import androidx.compose.ui.text.input.TextFieldValue
-import com.processout.sdk.api.model.response.POCard
-import com.processout.sdk.core.ProcessOutResult
 import com.processout.sdk.ui.core.state.POActionState
 import com.processout.sdk.ui.core.state.POFieldState
 import com.processout.sdk.ui.core.state.POImmutableList
@@ -32,17 +29,4 @@ internal data class CardTokenizationViewModelState(
         data class DropdownField(val state: POFieldState) : Item
         data class Group(val items: POImmutableList<Item>) : Item
     }
-}
-
-internal sealed interface CardTokenizationEvent {
-    data class FieldValueChanged(val id: String, val value: TextFieldValue) : CardTokenizationEvent
-    data class FieldFocusChanged(val id: String, val isFocused: Boolean) : CardTokenizationEvent
-    data class Action(val id: String) : CardTokenizationEvent
-    data class Dismiss(val failure: ProcessOutResult.Failure) : CardTokenizationEvent
-}
-
-internal sealed interface CardTokenizationCompletion {
-    data object Awaiting : CardTokenizationCompletion
-    data class Success(val card: POCard) : CardTokenizationCompletion
-    data class Failure(val failure: ProcessOutResult.Failure) : CardTokenizationCompletion
 }
